samba_dnsupdate: Do not rewrite krb5.conf in selftest
The samba_dnsupdate script provisions the DNS entries. The private
krb5.conf uses dns lookups to find the KDC to acquire a Kerberos ticket.
Obviously this will fail because are in the process of adding the DNS
entries for the KDC.

If we are inside of selftest we need to use the krb5.conf created by
selftest itself.

Signed-off-by: Andreas Schneider <asn@samba.org>