Allow py_smb_savefile() to except utf text
py_smb_loadfile() is able to read utf-16, but py_smb_savefile() was not able to write
utf-16. Using the 's#' string format allows an invisible length parameter to be passed
to the c code by python, those eliminating the need to do a strlen(), and making
the function compatible with other string encodings.

Signed-off-by: David Mulder <dmulder@suse.com>